backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
google-genai==1.1.0
google-adk==0.3.0
google-cloud-aiplatform==1.78.0
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
python-dotenv==1.0.1
